Mute: Restore Mute functionality through RIL for legacy devices [2/2]
This restores the old Mute function for devices that do not support
muting the microphone through the system audio manager.

Disable the Media Button callback, and send a HEADSETHOOK
Keyevent when the in-call mute button is pressed.

Issues with this method are that the in-call microphone button never gets
highlighted when we are in mute - but the statusbar icon will indicate whether
we are in a mute state.

add new build.prop entry to enable this: ro.telephony.legacy_mute=1

(See Telephony for part 1/2)

Good enough for now....

Change-Id: I873fb8cbc9d2e37e601c005450d0078d5fc8cab1